gpu: quiet CUDA header pedantic warnings in poker_gpu_unified and fix a double conversion
cuda_runtime.h's driver_types.h uses C11 anonymous unions, which the
project-wide -pedantic/-Werror build rejects - the same reason the CUDA
libraries already pass -Wno-error=pedantic. Apply it to poker_gpu_unified
too, and cast the size_t totalGlobalMem to double before dividing, which
-Wconversion flags.

+++ src/gpu/eval_batched_cuda.c
@@ -453,7 +453,7 @@ int cuda_backend_init(void** out_context, int device_id, bool verbose) {
         fprintf(stderr, "  Compute capability: %d.%d\n",
                 prop.major, prop.minor);
         fprintf(stderr, "  Global memory: %.2f GB\n",
-                prop.totalGlobalMem / (1024.0 * 1024.0 * 1024.0));
+                (double)prop.totalGlobalMem / (1024.0 * 1024.0 * 1024.0));
         fprintf(stderr, "  Lookup tables loaded to constant memory\n");
     }
